requirements.txt
pandas
jupyter
numpy
pillow
scikit-learn
openpyxl
